pyproject.toml
[project]
name = "silicon-empire"
version = "0.1.0"
description = "AI-Native Federated E-Commerce Organization"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # === LangGraph Core ===
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    
    # === MCP Integration ===
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    
    # === Database (本地 PostgreSQL) ===
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.3.0",
    
    # === Redis ===
    "redis[hiredis]>=5.0.0",
    
    # === Web/RPA ===
    "playwright>=1.40.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    
    # === Feishu SDK ===
    "lark-oapi>=1.3.0",
    
    # === Utilities ===
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "uuid7>=0.1.0",
    
    # === Observability ===
    "langsmith>=0.1.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
